+++ src/main/java/racingcar/domain/random/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.game;
+package racingcar.domain.random;
 
 public interface RandomNumberGenerator {
