+++ src/main/java/racingcar/exception/GetWinnerBeforeFinishException.java
@@ -1,4 +1,4 @@
-package racingcar.domain.exception;
+package racingcar.exception;
 
 public class GetWinnerBeforeFinishException extends IllegalStateException {
