+++ src/main/java/racingcar/exception/TryCountException.java
@@ -1,4 +1,4 @@
-package racingcar.parser.exception;
+package racingcar.exception;
 
 public class TryCountException extends IllegalArgumentException{
     public TryCountException(String message) {
