+++ src/main/java/racingcar/util/RetryableTemplate.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import java.util.function.Consumer;
