+++ src/test/java/racingcar/parser/TryCountParserTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.parser.exception.TryCountException;
+import racingcar.exception.TryCountException;
 
 public class TryCountParserTest {
